+++ tests/Imbo/UnitTest/EventListener/ImageTransformationCacheTest.php
@@ -11,7 +11,8 @@
 namespace Imbo\UnitTest\EventListener;
 
 use Imbo\EventListener\ImageTransformationCache,
-    org\bovigo\vfs\vfsStream;
+    org\bovigo\vfs\vfsStream,
+    org\bovigo\vfs\vfsStreamDirectory;
 
 /**
  * @author Christer Edvartsen <cogo@starzinger.net>
@@ -37,6 +38,7 @@ class ImageTransformationCacheTest extends ListenerTests {
     private $publicKey = 'publicKey';
     private $imageIdentifier = '7bf2e67f09de203da740a86cd37bbe8d';
     private $responseHeaders;
+    private $requestHeaders;
     private $cacheDir;
 
     /**
@@ -48,6 +50,7 @@ class ImageTransformationCacheTest extends ListenerTests {
         }
 
         $this->responseHeaders = $this->getMock('Imbo\Http\HeaderContainer');
+        $this->requestHeaders = $this->getMock('Imbo\Http\HeaderContainer');
         $this->query = $this->getMockBuilder('Imbo\Http\ParameterContainer')->disableOriginalConstructor()->getMock();
         $this->response = $this->getMock('Imbo\Http\Response\ResponseInterface');
         $this->response->expects($this->any())->method('getHeaders')->will($this->returnValue($this->responseHeaders));
@@ -56,6 +59,7 @@ class ImageTransformationCacheTest extends ListenerTests {
         $this->request->expects($this->any())->method('getQuery')->will($this->returnValue($this->query));
         $this->request->expects($this->any())->method('getPublicKey')->will($this->returnValue($this->publicKey));
         $this->request->expects($this->any())->method('getImageIdentifier')->will($this->returnValue($this->imageIdentifier));
+        $this->request->expects($this->any())->method('getHeaders')->will($this->returnValue($this->requestHeaders));
         $this->event->expects($this->any())->method('getRequest')->will($this->returnValue($this->request));
         $this->event->expects($this->any())->method('getResponse')->will($this->returnValue($this->response));
 
@@ -84,68 +88,69 @@ class ImageTransformationCacheTest extends ListenerTests {
 
     /**
      * @covers Imbo\EventListener\ImageTransformationCache::loadFromCache
-     */
-    public function testDoesNotLookInCacheIfNoTransformationsExists() {
-        $image = $this->getMock('Imbo\Model\Image');
-        $this->query->expects($this->once())->method('get')->with('t')->will($this->returnValue(null));
-        $this->response->expects($this->never())->method('getImage');
-
-        $this->listener->loadFromCache($this->event);
-    }
-
-    /**
-     * @covers Imbo\EventListener\ImageTransformationCache::loadFromCache
+     * @covers Imbo\EventListener\ImageTransformationCache::getCacheDir
      * @covers Imbo\EventListener\ImageTransformationCache::getCacheKey
      * @covers Imbo\EventListener\ImageTransformationCache::getCacheFilePath
      */
     public function testChangesTheImageInstanceOnCacheHit() {
         $imageFromCache = $this->getMock('Imbo\Model\Image');
-        $image = $this->getMock('Imbo\Model\Image');
+        $headersFromCache = $this->getMock('Imbo\Http\HeaderContainer');
+        $cachedData = serialize(array(
+            'image' => $imageFromCache,
+            'headers' => $headersFromCache,
+        ));
+
+        $this->request->expects($this->any())->method('getPublicKey')->will($this->returnValue($this->publicKey));
+        $this->request->expects($this->any())->method('getImageIdentifier')->will($this->returnValue($this->imageIdentifier));
+        $this->request->expects($this->any())->method('getExtension')->will($this->returnValue('png'));
+        $this->requestHeaders->expects($this->once())
+                             ->method('get')
+                             ->with('Accept', '*/*')
+                             ->will($this->returnValue(
+                                 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8'
+                             ));
 
         $this->query->expects($this->once())->method('get')->with('t')->will($this->returnValue(array('thumbnail')));
-        $this->response->expects($this->once())->method('getImage')->will($this->returnValue($image));
-        $this->response->expects($this->once())->method('setImage')->with($imageFromCache);
 
-        $this->request->expects($this->once())->method('getPublicKey')->will($this->returnValue($this->publicKey));
-        $this->request->expects($this->once())->method('getImageIdentifier')->will($this->returnValue($this->imageIdentifier));
-
-        $this->responseHeaders->expects($this->once())->method('set')->with('X-Imbo-TransformationCache', 'Hit');
+        $this->response->expects($this->once())->method('setModel')->with($imageFromCache)->will($this->returnSelf());
+        $this->response->expects($this->once())->method('setHeaders')->with($headersFromCache)->will($this->returnSelf());;
         $this->event->expects($this->once())->method('stopPropagation')->with(true);
 
-        $dir = 'vfs://cacheDir/p/u/b/publicKey/7/b/f/7bf2e67f09de203da740a86cd37bbe8d/c/9/5';
-        $file = 'c955032a862e0c8e76c8e51f5f402a6cf4e908ce013027e4a1400bc8754dcadd';
+        $dir = 'vfs://cacheDir/p/u/b/publicKey/7/b/f/7bf2e67f09de203da740a86cd37bbe8d/6/7/7';
+        $file = '677605632e7a57c58734e0a60cc1aaa7';
         $fullPath = $dir . '/' . $file;
 
         mkdir($dir, 0775, true);
-        file_put_contents($fullPath, serialize($imageFromCache));
+        file_put_contents($fullPath, $cachedData);
 
         $this->listener->loadFromCache($this->event);
     }
 
     /**
      * @covers Imbo\EventListener\ImageTransformationCache::loadFromCache
+     * @covers Imbo\EventListener\ImageTransformationCache::getCacheDir
      * @covers Imbo\EventListener\ImageTransformationCache::getCacheKey
      * @covers Imbo\EventListener\ImageTransformationCache::getCacheFilePath
      */
     public function testRemovesCorruptCachedDataOnCacheHit() {
-        $imageFromCache = $this->getMock('Imbo\Model\Image');
-        $image = $this->getMock('Imbo\Model\Image');
+        $this->request->expects($this->any())->method('getPublicKey')->will($this->returnValue($this->publicKey));
+        $this->request->expects($this->any())->method('getImageIdentifier')->will($this->returnValue($this->imageIdentifier));
+        $this->request->expects($this->any())->method('getExtension')->will($this->returnValue('png'));
+        $this->requestHeaders->expects($this->once())
+                             ->method('get')
+                             ->with('Accept', '*/*')
+                             ->will($this->returnValue(
+                                 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8'
+                             ));
 
         $this->query->expects($this->once())->method('get')->with('t')->will($this->returnValue(array('thumbnail')));
-        $this->response->expects($this->once())->method('getImage')->will($this->returnValue($image));
-        $this->response->expects($this->never())->method('setImage');
 
-        $this->request->expects($this->once())->method('getPublicKey')->will($this->returnValue($this->publicKey));
-        $this->request->expects($this->once())->method('getImageIdentifier')->will($this->returnValue($this->imageIdentifier));
-
-        $this->responseHeaders->expects($this->once())->method('set')->with('X-Imbo-TransformationCache', 'Miss');
-
-        $dir = 'vfs://cacheDir/p/u/b/publicKey/7/b/f/7bf2e67f09de203da740a86cd37bbe8d/c/9/5';
-        $file = 'c955032a862e0c8e76c8e51f5f402a6cf4e908ce013027e4a1400bc8754dcadd';
+        $dir = 'vfs://cacheDir/p/u/b/publicKey/7/b/f/7bf2e67f09de203da740a86cd37bbe8d/6/7/7';
+        $file = '677605632e7a57c58734e0a60cc1aaa7';
         $fullPath = $dir . '/' . $file;
 
         mkdir($dir, 0775, true);
-        file_put_contents($fullPath, 'invalid data');
+        file_put_contents($fullPath, 'corrupt data');
 
         $this->assertTrue(file_exists($fullPath));
         $this->listener->loadFromCache($this->event);
@@ -154,55 +159,53 @@ class ImageTransformationCacheTest extends ListenerTests {
 
     /**
      * @covers Imbo\EventListener\ImageTransformationCache::loadFromCache
+     * @covers Imbo\EventListener\ImageTransformationCache::getCacheDir
      * @covers Imbo\EventListener\ImageTransformationCache::getCacheKey
      * @covers Imbo\EventListener\ImageTransformationCache::getCacheFilePath
      */
     public function testAddsCorrectResponseHeaderOnCacheMiss() {
-        $imageFromCache = $this->getMock('Imbo\Model\Image');
-        $image = $this->getMock('Imbo\Model\Image');
-
-        $this->query->expects($this->once())->method('get')->with('t')->will($this->returnValue(array('thumbnail')));
-        $this->response->expects($this->once())->method('getImage')->will($this->returnValue($image));
-
-        $this->request->expects($this->once())->method('getPublicKey')->will($this->returnValue($this->publicKey));
-        $this->request->expects($this->once())->method('getImageIdentifier')->will($this->returnValue($this->imageIdentifier));
-
+        $this->requestHeaders->expects($this->once())
+                             ->method('get')
+                             ->with('Accept', '*/*')
+                             ->will($this->returnValue('*/*'));
         $this->responseHeaders->expects($this->once())->method('set')->with('X-Imbo-TransformationCache', 'Miss');
-
         $this->listener->loadFromCache($this->event);
     }
 
     /**
      * @covers Imbo\EventListener\ImageTransformationCache::storeInCache
      */
-    public function testDoesNotStoreImageInCacheIfNoTransformationHaveBeenApplied() {
-        $image = $this->getMock('Imbo\Model\Image');
-        $image->expects($this->once())->method('hasBeenTransformed')->will($this->returnValue(false));
-        $this->response->expects($this->once())->method('getImage')->will($this->returnValue($image));
-        $this->event->expects($this->never())->method('getRequest');
-
+    public function testDoesNotStoreNonImageModelsInTheCache() {
+        $this->response->expects($this->once())->method('getModel')->will($this->returnValue($this->getMock('Imbo\Model\Error')));
+        $this->request->expects($this->never())->method('getPublicKey');
         $this->listener->storeInCache($this->event);
     }
 
     /**
      * @covers Imbo\EventListener\ImageTransformationCache::storeInCache
+     * @covers Imbo\EventListener\ImageTransformationCache::getCacheDir
      * @covers Imbo\EventListener\ImageTransformationCache::getCacheKey
      * @covers Imbo\EventListener\ImageTransformationCache::getCacheFilePath
      */
-    public function testStoresImageInCacheWhenTransformationsHaveBeenApplied() {
+    public function testStoresImageInCache() {
         $image = $this->getMock('Imbo\Model\Image');
-        $image->expects($this->once())->method('hasBeenTransformed')->will($this->returnValue(true));
-        $fileContents = serialize($image);
 
-        $this->response->expects($this->once())->method('getImage')->will($this->returnValue($image));
-        $this->query->expects($this->once())->method('get')->with('t')->will($this->returnValue(array('thumbnail')));
+        $this->response->expects($this->once())->method('getModel')->will($this->returnValue($this->getMock('Imbo\Model\Image')));
+        $this->requestHeaders->expects($this->once())
+                             ->method('get')
+                             ->with('Accept', '*/*')
+                             ->will($this->returnValue('*/*'));
 
-        $cacheFile = 'vfs://cacheDir/p/u/b/publicKey/7/b/f/7bf2e67f09de203da740a86cd37bbe8d/c/9/5/c955032a862e0c8e76c8e51f5f402a6cf4e908ce013027e4a1400bc8754dcadd';
+        $cacheFile = 'vfs://cacheDir/p/u/b/publicKey/7/b/f/7bf2e67f09de203da740a86cd37bbe8d/5/a/a/5aa83f9df03e31b07c97299b927c6fd7';
 
         $this->assertFalse(is_file($cacheFile));
         $this->listener->storeInCache($this->event);
         $this->assertTrue(is_file($cacheFile));
-        $this->assertEquals($image, unserialize(file_get_contents($cacheFile)));
+
+        $data = unserialize(file_get_contents($cacheFile));
+
+        $this->assertEquals($image, $data['image']);
+        $this->assertEquals($this->responseHeaders, $data['headers']);
     }
 
     /**
@@ -211,17 +214,37 @@ class ImageTransformationCacheTest extends ListenerTests {
      * @covers Imbo\EventListener\ImageTransformationCache::rmdir
      */
     public function testCanDeleteAllImageVariationsFromCache() {
-        $dir = 'vfs://cacheDir/p/u/b/publicKey/7/b/f/7bf2e67f09de203da740a86cd37bbe8d/3/0/f';
-        $file = '30f0763c8422360d10fd84573dd582933a463e084b4f12b2b88eb1467e9eb338';
-        $fullPath = $dir . '/' . $file;
+        $cachedFiles = array(
+            'vfs://cacheDir/p/u/b/publicKey/7/b/f/7bf2e67f09de203da740a86cd37bbe8d/3/0/f/30f0763c8422360d10fd84573dd58293',
+            'vfs://cacheDir/p/u/b/publicKey/7/b/f/7bf2e67f09de203da740a86cd37bbe8d/3/0/e/30e0763c8422360d10fd84573dd58293',
+            'vfs://cacheDir/p/u/b/publicKey/7/b/f/7bf2e67f09de203da740a86cd37bbe8d/3/0/d/30d0763c8422360d10fd84573dd58293',
+        );
 
-        mkdir($dir, 0775, true);
-        file_put_contents($fullPath, 'image data');
+        foreach ($cachedFiles as $file) {
+            @mkdir(dirname($file), 0775, true);
+            file_put_contents($file, 'image data');
+            $this->assertTrue(is_file($file));
+        }
 
-        $this->assertTrue(is_file($fullPath));
         $this->listener->deleteFromCache($this->event);
-        $this->assertFalse(is_file($fullPath));
+
+        foreach ($cachedFiles as $file) {
+            $this->assertFalse(is_file($file));
+        }
+
         $this->assertFalse(is_dir('vfs://cacheDir/p/u/b/publicKey/7/b/f/7bf2e67f09de203da740a86cd37bbe8d'));
         $this->assertTrue(is_dir('vfs://cacheDir/p/u/b/publicKey/7/b/f'));
+    }
+
+    /**
+     * @expectedException PHPUnit_Framework_Error_Warning
+     * @expectedExceptionMessage Cache path is not writable by the webserver
+     * @covers Imbo\EventListener\ImageTransformationCache::__construct
+     */
+    public function testTriggersWarningIfCacheDirIsNotWritable() {
+        $dir = new vfsStreamDirectory('dir', 0);
+        $this->cacheDir->addChild($dir);
+
+        $listener = new ImageTransformationCache('vfs://cacheDir/dir');
     }
 }